+++ lib/pusher_client_socket.dart
@@ -55,7 +55,9 @@ class PusherClient {
 
   /// Connects the client to the server.
   void connect() {
-    __socket = WebSocket(Uri.parse(options.url));
+    __socket = WebSocket(
+      options.uri,
+    );
 
     _socket.connection.listen(_onConnectionStateChange);
 
@@ -99,6 +101,7 @@ class PusherClient {
       case Connected():
         _onEvent('connected', null);
         _resetActivityCheck();
+        _reconnectionAttempts = 0;
         break;
       case Reconnecting():
         _onEvent('reconnecting', null);
@@ -106,6 +109,7 @@ class PusherClient {
       case Reconnected():
         _onEvent('reconnected', null);
         _resetActivityCheck();
+        _reconnectionAttempts = 0;
         break;
       case Disconnecting():
         _onEvent('disconnecting', null);
@@ -129,8 +133,19 @@ class PusherClient {
 
     disconnect(1006, error.message);
 
-    if (error is WebSocketException) {
+    if (error is SocketException) {
+      _reconnect();
+    }
+  }
+
+  int _reconnectionAttempts = 0;
+  void _reconnect() async {
+    if (_reconnectionAttempts < options.maxReconnectionAttempts) {
+      _reconnectionAttempts++;
+      await Future.delayed(options.reconnectGap);
       connect();
+    } else {
+      disconnect(null, "Max reconnection attempts reached");
     }
   }